Keycloak security

+++ common/common-model/src/main/java/com/evo/common/dto/response/ProductDTO.java
@@ -6,6 +6,7 @@ import java.util.Map;
 import java.util.UUID;
 
 import com.evo.common.enums.DiscountType;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -23,12 +24,12 @@ public class ProductDTO {
     private UUID categoryId;
     private Map<String, String> description;
     private String introduce; // Introduce is a short description of the product(html)
-    private Double weight;
-    private Double length;
-    private Double width;
+    private int weight;
+    private int length;
+    private int width;
     private Long totalSold;
     private BigDecimal averageRating;
-    private Double height;
+    private int height;
     private Boolean hidden;
     List<ProductVariantDTO> productVariants;
     List<ProductImageDTO> productImages;
